+++ packages/babel-preset/utils.spec.js
@@ -1,4 +1,3 @@
-/* eslint-disable unicorn/prefer-module */
 const { isDevelopment } = require('./utils');
 
 describe('utils', () => {