+++ packages/babel-preset/web/index.js
@@ -1,8 +1,6 @@
-/* eslint-disable unicorn/prefer-module */
 const sharedPlugins = require('../sharedPlugins');
 const { isDevelopment, isDebugging } = require('../utils');
 
-// eslint-disable-next-line unicorn/prefer-module
 module.exports = function autoGuruWebPreset(api, options = {}) {
 	const { modules = false, debug = false, corejs = 3, browsers } = options;