+++ packages/browserslist-config/test.spec.js
@@ -1,4 +1,3 @@
-/* eslint-disable unicorn/prefer-module */
 const browserslist = require('browserslist');
 
 const config = require('.');