+++ packages/gdu/config/babel.config.js
@@ -1,9 +1,14 @@
-/* eslint-disable unicorn/prefer-module */
 const { join } = require('path');
 
 const browsers = require('browserslist-config-autoguru');
 
 const { PROJECT_ROOT } = require('../lib/roots');
+const ReactCompilerConfig = {
+	target: '19',
+	sources: (filename) => {
+		return filename.includes('apps') || filename.includes('packages');
+	},
+};
 
 module.exports = (guruConfig) => {
 	let hasRelay = false;
@@ -37,7 +42,7 @@ module.exports = (guruConfig) => {
 				require.resolve('@autoguru/babel-preset/react'),
 			],
 			[
-				require.resolve('@babel/preset-typescript'),
+				'@babel/preset-typescript',
 				{
 					isTSX: true,
 					allExtensions: true,
@@ -45,8 +50,9 @@ module.exports = (guruConfig) => {
 			],
 		].filter(Boolean),
 		plugins: [
+			['babel-plugin-react-compiler', ReactCompilerConfig],
 			hasRelay && [
-				require.resolve('babel-plugin-relay'),
+				'babel-plugin-relay',
 				{
 					haste: false,
 					isDevVariable: '__DEV__',