+++ packages/gdu/lib/configure.ts
@@ -78,9 +78,7 @@ export const configure = async (app_location) => {
 		},
 		include: [
 			'./*.d.ts',
-			...guruConfig.srcPaths.map(
-				(item) => `${item.replace(/\/$/, '')}/**/*`,
-			),
+			...guruConfig.srcPaths.map((item) => `${item.replace(/\/$/, '')}`),
 		].filter(Boolean),
 		exclude: ['node_modules'],
 	};