+++ packages/gdu/lib/roots.ts
@@ -2,7 +2,7 @@ import { join } from 'path';
 
 import { diary } from 'diary';
 import getYarnWorkspaceRoot from 'find-yarn-workspace-root';
-/* eslint-disable unicorn/prefer-module */
+
 const { debug } = diary('gdu:roots');
 
 export const GDU_ROOT = join(__dirname, '../../');