+++ packages/utilities/jest.config.js
@@ -1,6 +1,4 @@
-/* eslint-disable unicorn/prefer-module */
 module.exports = {
-	// eslint-disable-next-line import/extensions
 	displayName: require('./package.json').name,
 	preset: '@autoguru/jest-preset',
 };