+++ packages/utilities/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @autoguru/utilities
 
+## 1.3.0
+
+### Minor Changes
+
+-   28011c4: Update Babel presets and configurations for React 19 support
+
 ## 1.2.2
 
 ### Patch Changes